Add Done Also

+++ OstManSysMVVM/Model/Apartment.cs
@@ -11,9 +11,9 @@ namespace OstManSysMVVM.Model
     {
         public int ApartmentID { get; set; }
         public string Address { get; set; }
-        public double Size { get; set; }
+        public int Size { get; set; }
         public int NumberOfRooms { get; set; }
-        public double MonthlyRent { get; set; }
+        public decimal MonthlyRent { get; set; }
         public string Condition { get; set; }
         public bool IsRented { get; set; }
         public DateTime LastCheck { get; set; }
